+++ samples/Sentry.Samples.Maui/AppShell.xaml.cs
@@ -5,5 +5,6 @@ public partial class AppShell
     public AppShell()
     {
         InitializeComponent();
+        Routing.RegisterRoute("ctmvvm", typeof(CtMvvmPage));
     }
 }